Migrate LineChartCard to recharts
