+++ package.json
@@ -74,7 +74,8 @@
     "react-external-link": "^2.2.0",
     "react-leaflet": "^4.2.1",
     "react-swipeable": "^7.0.1",
-    "react-tag-autocomplete": "^7.1.0"
+    "react-tag-autocomplete": "^7.1.0",
+    "recharts": "^2.10.3"
   },
   "devDependencies": {
     "@shlinkio/eslint-config-js-coding-standard": "~2.3.0",
