+++ src/visits/VisitsStats.tsx
@@ -15,12 +15,12 @@ import { toDateRange } from '../utils/dates/helpers/dateIntervals';
 import { prettify } from '../utils/helpers/numbers';
 import { useSetting } from '../utils/settings';
 import { DoughnutChartCard } from './charts/DoughnutChartCard';
-import { LineChartCard } from './charts/LineChartCard';
 import { SortableBarChartCard } from './charts/SortableBarChartCard';
 import { useVisitsQuery } from './helpers/hooks';
 import { OpenMapModalBtn } from './helpers/OpenMapModalBtn';
 import { VisitsFilterDropdown } from './helpers/VisitsFilterDropdown';
 import { VisitsStatsOptions } from './helpers/VisitsStatsOptions';
+import { LineChartCard } from './recharts/LineChartCard';
 import type { VisitsDeletion, VisitsInfo } from './reducers/types';
 import { normalizeVisits, processStatsFromVisits } from './services/VisitsParser';
 import type { NormalizedOrphanVisit, NormalizedVisit, VisitsParams } from './types';
