--- src/visits/charts/LineChartCard.scss
@@ -1,9 +1,0 @@
-@import '../../../node_modules/@shlinkio/shlink-frontend-kit/dist/base';
-
-.line-chart-card__body canvas {
-  height: 300px !important;
-
-  @media (min-width: $mdMin) {
-    height: 400px !important;
-  }
-}
